Improve wiring of the build info task
Before this change the task was hardcoding dependencies
to `compileJava` and friends. It was also messing with the
default classes directory, which is not great because it
kills caching.

Instead, this commit wires the task so that it generates
its output in its own directory, which is instead added
as a generated resource.

The advantage is that all wiring is implicit: all tasks
which require the resources will automatically trigger
the creation of the build info file (in particular `jar`).
It's also going to be implicitly added to the classpath
for tests for example.

Eventually, the task now properly declares all its inputs,
making it cacheable and compatible with the configuration
cache.